Ajout des Listener avec Classe EcouteSoldats
liaison des positions des soldats et de leurs sprites avec HashMap

Opti de l'affichage des soldats

+++ src/main/java/fr/iut/montreuil/Red_Line_Defense/modele/ActeursJeu/Rookie.java
@@ -2,7 +2,7 @@ package fr.iut.montreuil.Red_Line_Defense.modele.ActeursJeu;
 
 public class Rookie extends Soldat {
 
-    public Rookie(double x0, double y0, double x1, double y1) {
+    public Rookie(int x0, int y0, int x1, int y1) {
         super(x0, y0, 10000, 300, 20, true, 250, x1, y1);       // Ennemi True car ce sera un Soldat exclusivement Ennemi
     }
 }
